8.2 - 1) added delete and create trips methods 2) removed mock фтв unused modules
