--- src/mock-data/trip-event-date-data.js
@@ -1,43 +1,0 @@
-import {getRandom} from '../utils/common';
-
-const DateSettings = {
-  MAX_INTERVAL_FOR_START_DAY: 15,
-  MIN_INTERVAL_FOR_END_DAY: 0.1,
-  MAX_INTERVAL_FOR_END_DAY: 0.2,
-  ONE_DAY_ON_MILLISECONDS: 24 * 3600 * 1000,
-};
-
-const getStartDate = () => {
-  if (Math.random() > 0.5) {
-    return Date.now() + (getRandom(DateSettings.MAX_INTERVAL_FOR_START_DAY) * DateSettings.ONE_DAY_ON_MILLISECONDS);
-  }
-  return Date.now() - (getRandom(DateSettings.MAX_INTERVAL_FOR_START_DAY) * DateSettings.ONE_DAY_ON_MILLISECONDS);
-};
-
-const getEndDate = (startDateValue) => startDateValue + (getRandom(DateSettings.MAX_INTERVAL_FOR_END_DAY, DateSettings.MIN_INTERVAL_FOR_END_DAY) * DateSettings.ONE_DAY_ON_MILLISECONDS);
-
-const generateTripEventDateData = () => {
-  const startDate = getStartDate();
-
-  return {
-    startDate: new Date(startDate),
-    endDate: new Date(getEndDate(startDate)),
-  };
-};
-
-const getTripDaysString = (item) => (`${item.date.startDate.getFullYear()}-${item.date.startDate.getMonth() + 1}-${item.date.startDate.getDate()}`);
-
-const generateTripDays = (eventArr) => {
-  let tripDays = eventArr
-    .slice()
-    .sort((a, b) => new Date(a.date.startDate) - new Date(b.date.startDate))
-    .map((item) => getTripDaysString(item));
-
-  return [...new Set(tripDays)];
-};
-
-export {
-  generateTripEventDateData,
-  generateTripDays,
-  getTripDaysString
-};
